+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.3 - ???
+==================
+
+- Bumped dependency to raven 2.0.6 and implied dependency on sentry 5.0.13
+- Added a DSN configuration to replace the project ID for the plugin
+
 0.2 - 2012-07-20
 ==================